requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
tensorflow==2.16.1
tensorflow-hub==0.16.1
numpy==1.26.4
opencv-python==4.9.0.80
Pillow==10.3.0